+++ src/main/java/kr/hhplus/be/server/presentation/user/dto/request/PointChargeRequest.java
@@ -4,6 +4,6 @@ import jakarta.validation.constraints.*;
 
 public record PointChargeRequest(
     @NotNull(message = "충전 금액은 필수입니다")
-    @Min(value = 1, message = "충전 금액은 1 이상이어야 합니다")
-    Integer amount
+    @Min(value = 1L, message = "충전 금액은 1 이상이어야 합니다")
+    Long amount
 ) {}
